+++ tests/components/supernotify/recipes/test_alarm_actions.py
@@ -80,8 +80,8 @@ async def test_mobile_push_only_has_arm_when_alarm_disarmed(fixture, hass: HomeA
     await uut.initialize()
     await uut.deliver()
     assert uut.selected_scenario_names == ["alarm_disarmed"]
-    assert len(uut.deliveries["apple_push"]["delivered_envelopes"]) == 1
-    envelope: Envelope = uut.deliveries["apple_push"]["delivered_envelopes"][0]  # type:ignore
+    assert len(uut.deliveries["apple_push"]["delivered"]) == 1
+    envelope: Envelope = uut.deliveries["apple_push"]["delivered"][0]  # type:ignore
     assert envelope.delivery_name == "apple_push"
     assert envelope.calls[0].action_data["data"]["actions"] == [  # type: ignore[index]
         {"action": "ALARM_PANEL_RESET", "title": "Arm Alarm Panel for at Home", "icon": "sfsymbols:bell"},
@@ -98,8 +98,8 @@ async def test_mobile_push_only_has_disarm_when_alarm_armed(fixture, hass: HomeA
     await uut.initialize()
     await uut.deliver()
     assert uut.selected_scenario_names == ["alarm_armed"]
-    assert len(uut.deliveries["apple_push"]["delivered_envelopes"]) == 1
-    envelope: Envelope = uut.deliveries["apple_push"]["delivered_envelopes"][0]  # type:ignore
+    assert len(uut.deliveries["apple_push"]["delivered"]) == 1
+    envelope: Envelope = uut.deliveries["apple_push"]["delivered"][0]  # type:ignore
     assert envelope.delivery_name == "apple_push"
     assert envelope.calls[0].action_data["data"]["actions"] == [  # type: ignore[index]
         {"action": "ALARM_PANEL_DISARM", "title": "Disarm Alarm Panel", "icon": "sfsymbols:bell.slash"},