+++ tests/components/supernotify/recipes/test_alexa_whisper.py
@@ -41,10 +41,10 @@ async def test_alexa_whispering(hass: HomeAssistant):
     uut = Notification(ctx, "testing 123", action_data={"priority": "low"}, target="joe@soapy.com")
     await uut.initialize()
     await uut.deliver()
-    assert len(uut.deliveries["alexa_inform"]["delivered_envelopes"]) == 1
-    assert len(uut.deliveries["plain_email"]["delivered_envelopes"]) == 1
-    assert uut.deliveries["plain_email"]["delivered_envelopes"][0].calls[0].action_data["message"] == "testing 123"  # type: ignore
+    assert len(uut.deliveries["alexa_inform"]["delivered"]) == 1
+    assert len(uut.deliveries["plain_email"]["delivered"]) == 1
+    assert uut.deliveries["plain_email"]["delivered"][0].calls[0].action_data["message"] == "testing 123"  # type: ignore
     assert (
-        uut.deliveries["alexa_inform"]["delivered_envelopes"][0].calls[0].action_data["message"]  # type: ignore
+        uut.deliveries["alexa_inform"]["delivered"][0].calls[0].action_data["message"]  # type: ignore
         == '<amazon:effect name="whispered">testing 123</amazon:effect>'
     )