+++ tests/components/supernotify/recipes/test_seasonal_scenario.py
@@ -72,9 +72,9 @@ async def test_seasonal_scenario_on_xmas_day(recipe_fixture):
     await uut.initialize()
     await uut.deliver()
 
-    assert len(uut.deliveries["doorbell_rang"]["delivered_envelopes"]) == 1
-    assert len(uut.deliveries["doorbell_rang"]["delivered_envelopes"][0].calls) == 1  # type:ignore
-    call = uut.deliveries["doorbell_rang"]["delivered_envelopes"][0].calls[0]  # type:ignore
+    assert len(uut.deliveries["doorbell_rang"]["delivered"]) == 1
+    assert len(uut.deliveries["doorbell_rang"]["delivered"][0].calls) == 1  # type:ignore
+    call = uut.deliveries["doorbell_rang"]["delivered"][0].calls[0]  # type:ignore
     assert call.action == "send_sound"
     assert call.domain == "alexa_devices"
     assert call.action_data["sound"] == "christmas_05"  # ty:ignore[non-subscriptable]
@@ -87,9 +87,9 @@ async def test_unseasonal_scenario_on_xmas_day(recipe_fixture):
     await uut.initialize()
     await uut.deliver()
 
-    assert len(uut.deliveries["driveway_alarm"]["delivered_envelopes"]) == 1
-    assert len(uut.deliveries["driveway_alarm"]["delivered_envelopes"][0].calls) == 1  # type:ignore
-    call = uut.deliveries["driveway_alarm"]["delivered_envelopes"][0].calls[0]  # type:ignore
+    assert len(uut.deliveries["driveway_alarm"]["delivered"]) == 1
+    assert len(uut.deliveries["driveway_alarm"]["delivered"][0].calls) == 1  # type:ignore
+    call = uut.deliveries["driveway_alarm"]["delivered"][0].calls[0]  # type:ignore
     assert call.action == "send_sound"
     assert call.domain == "alexa_devices"
     assert call.action_data["sound"] == "amzn_sfx_trumpet_bugle_04"  # ty:ignore[non-subscriptable]
@@ -102,9 +102,9 @@ async def test_seasonal_scenario_in_summer(recipe_fixture):
     await uut.initialize()
     await uut.deliver()
 
-    assert len(uut.deliveries["doorbell_rang"]["delivered_envelopes"]) == 1
-    assert len(uut.deliveries["doorbell_rang"]["delivered_envelopes"][0].calls) == 1  # type:ignore
-    call = uut.deliveries["doorbell_rang"]["delivered_envelopes"][0].calls[0]  # type:ignore
+    assert len(uut.deliveries["doorbell_rang"]["delivered"]) == 1
+    assert len(uut.deliveries["doorbell_rang"]["delivered"][0].calls) == 1  # type:ignore
+    call = uut.deliveries["doorbell_rang"]["delivered"][0].calls[0]  # type:ignore
     assert call.action == "send_sound"
     assert call.domain == "alexa_devices"
     assert call.action_data["sound"] == "amzn_sfx_doorbell_chime_02"  # ty:ignore[non-subscriptable]