+++ tests/components/supernotify/test_archive.py
@@ -60,7 +60,7 @@ async def test_integration_archive(mock_hass: HomeAssistant, debug: bool) -> Non
             blob: str = "".join(await stream.readlines())
             reobj = json.loads(blob)
         assert reobj["priority"] == "critical"
-        for outcome in ("delivered_envelopes", "delivered_envelopes", "no_envelopes"):
+        for outcome in ("delivered", "failed", "skipped", "no_envelopes"):
             for delivery_name in uut.last_notification.deliveries:
                 assert len(reobj["deliveries"][delivery_name].get(outcome, [])) == len(
                     uut.last_notification.deliveries[delivery_name].get(outcome, [])