+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### New
 - Scenarios can now use regular expressions for the delivery configuration, for example `.*` to apply to every one
 - `enquire_deliveries_by_scenario` action now lists which deliveries are enabled by the scenario, disabled by it, and all the scenarios to which it applies overrides
+- Further improvements to `Notification` archive object to make it easier to debug
+  - Original message now at top of object
+  - Envelopes for each delivery categorized into `delivered`,`failed`,`skipped` and `no_envelopes`
 ## 1.6.0
 ### Fixes
 - Image attachments from mobile push style notifications not being picked up by e-mail